backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0        
celery==5.1.2
certifi==2021.10.8       
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0  
click-plugins==1.1.1     
click-repl==0.2.0        
dj-database-url==0.5.0   
Django==3.2.8
django-anymail==8.4      
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-environ==0.8.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
idna==3.3
iniconfig==1.1.1
kombu==5.1.0
mysqlclient==2.0.3
packaging==21.2
Pillow==8.4.0
pluggy==1.0.0
prompt-toolkit==3.0.21
py==1.10.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytest-mock==3.6.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5